Moved LWJGL object code to lwjgl project
