+++ jgloom/src/main/java/jgloom/GLNativeException.java
@@ -1,7 +1,5 @@
 package jgloom;
 
-import org.lwjgl.opengl.GL11;
-
 /**
  * Signals that an OpenGL method has been invoked / used at an incorrect
  * or inappropriate time.
@@ -24,14 +22,5 @@ public class GLNativeException extends IllegalStateException {
      */
     public GLNativeException(int error){
         super(""+error);
-    }
-
-    /**
-     * Checks for and throws OpenGL errors
-     * @throws GLNativeException If an OpenGL error exists
-     */
-    public static void checkOGL() throws GLNativeException {
-        if(GL11.glGetError() != GL11.GL_NO_ERROR)
-            throw new GLNativeException(GL11.glGetError());
     }
 }
