+++ jgloom/src/main/java/jgloom/gl/GLBuffer.java
@@ -1,8 +1,5 @@
 package jgloom.gl;
 
-import org.lwjgl.opengl.GL15;
-import org.lwjgl.opengl.GL43;
-
 /**
  * Buffer Objects are OpenGL Objects that store an array of unformatted memory allocated by the OpenGL context (aka:
  * the GPU). These can be used to store vertex data, pixel data retrieved from images or the framebuffer, and a variety
@@ -13,23 +10,4 @@ import org.lwjgl.opengl.GL43;
 public interface GLBuffer {
     /** @return The identifier for the buffer object */
     int getBuffer();
-
-    /** The OpenGL framebuffer object identifier */
-    int IDENTIFIER = GL43.GL_BUFFER;
-
-    /**
-     * Constructs a {@link GLBuffer} object with {@link GL15#glGenBuffers()}.
-     */
-    static GLBuffer createBuffer() {
-        int buffer = GL15.glGenBuffers();
-        return () -> buffer;
-    }
-
-    /**
-     * @param buffer
-     * @return If an integer corresponds to an OpenGL buffer object
-     */
-    static boolean isBuffer(int buffer){
-        return GL15.glIsBuffer(buffer);
-    }
 }
